+++ src/uqtestfuns/core/__init__.py
@@ -3,13 +3,13 @@ The core subpackage of uqtestfuns.
 """
 
 from .parameters import FunParams
-from .prob_input.univariate_distribution import UnivDist
+from .prob_input.marginal import Marginal
 from .prob_input.probabilistic_input import ProbInput
 from .uqtestfun_abc import UQTestFunBareABC, UQTestFunABC
 from .uqtestfun import UQTestFun
 
 __all__ = [
-    "UnivDist",
+    "Marginal",
     "ProbInput",
     "FunParams",
     "UQTestFunBareABC",